Change max value

+++ src/feeder/terminal.rs
@@ -164,7 +164,7 @@ impl Terminal {
     }
 
     pub fn call_history(&mut self, inc: i32, core: &mut ShellCore){
-        Self::shift_in_range(&mut self.hist_ptr, inc, 0, std::usize::MAX);
+        Self::shift_in_range(&mut self.hist_ptr, inc, 0, std::i32::MAX as usize);
 
         self.chars = self.prompt.chars().collect();
         self.chars.extend(core.fetch_history_file(self.hist_ptr).chars());
